+++ lib/test/clojure_lsp/feature/clean_ns_test.clj
@@ -115,7 +115,7 @@
                             "     java.util.Date))"
                             "File Date")))))
   (testing "with first require as unused"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [foo  :as f] [bar :as b] baz [z] ))"
@@ -129,7 +129,7 @@
                            "(defn func []"
                            "  (b/some))")))
   (testing "with single unused require on ns"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [foo  :as f] ))"
@@ -139,7 +139,7 @@
                            "(defn func []"
                            "  (b/some))")))
   (testing "with single used require on ns"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [foo  :as f] ))"
@@ -151,14 +151,14 @@
                            "(defn func []"
                            "  (f/some))")))
   (testing "with multiple unused requires on ns"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [foo  :as f]"
                            "   [bar :as b]))")
                    (h/code "(ns foo.bar)")))
   (testing "with duplicate require with different and unused alias"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "  [bar :as b]"
@@ -171,7 +171,7 @@
                            "  [foo :as f]))"
                            "f/bar b/bar")))
   (testing "with duplicate require with both used alias"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "  [bar :as b]"
@@ -185,7 +185,7 @@
                            "  [foo :as fa]))"
                            "f/bar fa/bar b/bar")))
   (testing "with refer at require"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [foo  :as f] [bar :refer [some]] baz [z] ))"
@@ -199,12 +199,12 @@
                            "(defn func []"
                            "  (f/some))")))
   (testing "with refer as single require"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [bar :refer [some]]))")
                    (h/code "(ns foo.bar)"))
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [bar :refer :all]))")
@@ -218,7 +218,8 @@
                            ""
                            "(defn func []"
                            "  (f/some))")]
-      (test-clean-ns {:documents {(h/file-uri "file:///a.clj") {:text to-clean}}}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}
+                      :documents {(h/file-uri "file:///a.clj") {:text to-clean}}}
                      to-clean
                      (h/code "(ns foo.bar"
                              " (:require"
@@ -230,7 +231,7 @@
                              "  (f/some))")
                      false)))
   (testing "with first require as a refer"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [bar :refer [some] ] [foo :as f]))"
@@ -244,7 +245,7 @@
                            "(defn func []"
                            "  (some))")))
   (testing "with first require as a refer with alias"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [bar :as b :refer [some] ] [foo :as f]))"
@@ -260,7 +261,7 @@
                            "  b/some"
                            "  (some))")))
   (testing "unused refer from multiple refers"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [bar :refer [some other] ]))"
@@ -270,7 +271,7 @@
                            "  [bar :refer [some] ]))"
                            "(some)")))
   (testing "unused refer and alias"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "  [baz]"
@@ -279,7 +280,7 @@
                            " (:require"
                            "  [baz]))")))
   (testing "unused refer from single refer but used alias before"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [aba :as a]"
@@ -293,7 +294,7 @@
                            "(a/bla)"
                            "(b/another)")))
   (testing "used refer from single refer and used alias after refer"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [aba :as a]"
@@ -309,7 +310,7 @@
                            "(b/another)"
                            "some")))
   (testing "unused refer from single refer but used alias after"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [aba :as a]"
@@ -323,7 +324,7 @@
                            "(a/bla)"
                            "(b/another)")))
   (testing "unused refer from multiple refers but used alias"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [bar :as b :refer [some other]]))"
@@ -335,7 +336,7 @@
                            "(other)"
                            "(b/another)")))
   (testing "unused middle refer from multiple refers"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [bar :refer [some other baz another] ]))"
@@ -349,7 +350,7 @@
                            "(another)"
                            "(baz)")))
   (testing "unused refer and alias"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:require"
                            "   [bar :refer [some] ]"
@@ -357,7 +358,7 @@
                    (h/code "(ns foo.bar)")))
   (testing "sorting"
     (testing "sorts according to symbols not brackets"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "  []"
@@ -470,7 +471,7 @@
                              "Foo"
                              "ball.")))
     (testing "unsorted used imports"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:import"
                              "  a.c.d.e.A"
@@ -544,7 +545,7 @@
                              "  (a.c E F G H I)))"
                              "  A B C D E F G H I")))
     (testing "unsorted used refer"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "   [some :refer [Dee foo bar baz]]))"
@@ -554,7 +555,8 @@
                              "  [some :refer [bar baz Dee foo]]))"
                              "   foo bar baz Dee")))
     (testing "unsorted used refer with less max-line-length"
-      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 30}}}}}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line
+                                         :sort {:refer {:max-line-length 30}}}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "   [some :refer [Dee foo bar baz bla blowning]]))"
@@ -565,7 +567,8 @@
                              "                blowning Dee"
                              "                foo]]))"
                              "   foo bar baz Dee bla blowning"))
-      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 30}}}}}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line
+                                         :sort {:refer {:max-line-length 30}}}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "  [some :refer [bar baz bla"
@@ -578,7 +581,8 @@
                              "                blowning Dee"
                              "                foo]]))"
                              "   foo bar baz Dee bla blowning"))
-      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 30}}}}}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line
+                                         :sort {:refer {:max-line-length 30}}}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "   [some :refer [Dee foo bar baz bla blowning cat car clown cobble doo did done danger fear fight]]))"
@@ -594,7 +598,8 @@
                              "                fear fight"
                              "                foo]]))"
                              "   foo bar baz Dee bla blowning cat car clown cobble doo did done danger fear fight"))
-      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 40}}}}}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line
+                                         :sort {:refer {:max-line-length 40}}}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "   [some :refer [Dee foo bar baz bla blowning] :as s]))"
@@ -606,7 +611,8 @@
                              "                Dee foo] :as s]))"
                              "   foo bar baz Dee bla blowning"
                              "   s/foo"))
-      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 40}}}}}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line
+                                         :sort {:refer {:max-line-length 40}}}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "   [some :as s :refer [Dee foo bar baz bla blowning]]))"
@@ -619,7 +625,8 @@
                              "   foo bar baz Dee bla blowning"
                              "   s/foo")))
     (testing "unsorted used refer with infinite max-line-length"
-      (test-clean-ns {:settings {:clean {:sort {:refer {:max-line-length 0}}}}}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line
+                                         :sort {:refer {:max-line-length 0}}}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "   [some :refer [Dee foo bar baz bla blowning]]))"
@@ -629,7 +636,8 @@
                              "  [some :refer [bar baz bla blowning Dee foo]]))"
                              "   foo bar baz Dee bla blowning")))
     (testing "unsorted used refer with sort disabled"
-      (test-clean-ns {:settings {:clean {:sort {:refer false}}}}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line
+                                         :sort {:refer false}}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "   [some :refer [foo bar baz]]))"
@@ -640,7 +648,7 @@
                              "   foo bar baz")))
     (testing "ns children sorting"
       (testing "keep comments"
-        (test-clean-ns {}
+        (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                        (h/code "(ns foo"
                                " (:require"
                                "  [clojure.set :as set] ;; important comment"
@@ -659,7 +667,7 @@
                                "set/a"
                                "edn/b")))
       (testing "import before require"
-        (test-clean-ns {}
+        (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                        (h/code "(ns foo.bar"
                                " (:import"
                                "  [foo Bar])"
@@ -675,7 +683,7 @@
                                "str/join"
                                "Bar")))
       (testing "import before require with other list between"
-        (test-clean-ns {}
+        (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                        (h/code "(ns foo.bar"
                                " (:import"
                                "  [foo Bar"
@@ -711,19 +719,19 @@
                                "str/join"
                                "Bar")))))
   (testing "single unused full package import"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:import"
                            "  java.util.Date))")
                    (h/code "(ns foo.bar)")))
   (testing "single unused package import"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:import"
                            "  [java.util Date]))")
                    (h/code "(ns foo.bar)")))
   (testing "unused full package imports"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:import "
                            "  java.util.Date java.util.Calendar java.util.List))"
@@ -733,7 +741,7 @@
                            "  java.util.Calendar))"
                            "Calendar.")))
   (testing "unused package imports"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:import "
                            "  [java.util Date Calendar List Map]))"
@@ -761,7 +769,7 @@
                            "Map."
                            "Foo. Bar.")))
   (testing "unused package imports with single import"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:import"
                            "  [java.util Date List]"
@@ -775,7 +783,7 @@
                            "Calendar."
                            "List.")))
   (testing "unused package imports spacing"
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:import"
                            "  [java.util Date"
@@ -788,7 +796,7 @@
                            "  [java.util Date List]))"
                            "Date."
                            "List."))
-    (test-clean-ns {}
+    (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                    (h/code "(ns foo.bar"
                            " (:import"
                            "  [java.util Date"
@@ -802,7 +810,7 @@
                            "List.")))
   (testing "cljc conditional readers"
     (testing "remove reader conditional after removing unused single require alias"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "  #?(:clj [other.zeta :as z])))")
@@ -810,7 +818,7 @@
                      true
                      "file:///a.cljc"))
     (testing "remove reader conditional after removing unused require alias"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "  #?(:cljs [other.foo :as o]
@@ -828,7 +836,7 @@
                      true
                      "file:///a.cljc"))
     (testing "remove reader conditional after removing unused require refer"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "  #?(:cljs [other.foo :refer [o]]
@@ -846,7 +854,7 @@
                      true
                      "file:///a.cljc"))
     (testing "remove reader conditional after removing unused import"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:import"
                              "  #?(:cljs [other.foo O]
@@ -864,7 +872,7 @@
                      true
                      "file:///a.cljc"))
     (testing "remove single unused import inside splicing reader conditional"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:import"
                              "  [java.util Calendar]"
@@ -879,7 +887,7 @@
                      true
                      "file:///a.cljc"))
     (testing "only used import in specific lang for splicing reader conditional"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:import"
                              "  [java.util Calendar]"
@@ -897,7 +905,7 @@
                      true
                      "file:///a.cljc"))
     (testing "only used required alias in specific lang"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "  [other.foo :as f]"
@@ -914,7 +922,7 @@
                      true
                      "file:///a.cljc"))
     (testing "only used required refer in specific lang"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:require"
                              "  [other.foo :refer [f]]"
@@ -931,7 +939,7 @@
                      true
                      "file:///a.cljc"))
     (testing "only used import in specific lang"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns foo.bar"
                              " (:import"
                              "  [other.foo F]"
@@ -948,7 +956,7 @@
                      true
                      "file:///a.cljc"))
     (testing "sort reader conditionals before normal requires"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns a"
                              "  (:require"
                              "   [a.b]"
@@ -970,7 +978,7 @@
                      true
                      "file:///a.cljc"))
     (testing "Mixed cases"
-      (test-clean-ns {}
+      (test-clean-ns {:settings {:clean {:ns-inner-blocks-indentation :next-line}}}
                      (h/code "(ns a"
                              "  (:require"
                              "   [c.b]"